remove: AbstractPathFinder 제거 및 PathFinder로 코드 이전
