+++ src/test/java/subway/acceptance/path/PathAcceptanceTest.java
@@ -178,6 +178,20 @@ public class PathAcceptanceTest extends AcceptanceTest {
     // week 4-2
 
     /**
+     * 1~10  : 1250
+     * 11~15 : 1350
+     * 16~20 : 1450
+     * 21~25 : 1550
+     * 26~30 : 1660
+     * 31~35 : 1750
+     * 36~40 : 1850
+     * 41~45 : 1950
+     * 46~50 : 2050
+     * 51~58 : 2150
+     * 59~.. : 2250
+     */
+
+    /**
      * Given 2개의 구간을 가진 노선이 있고
      * When 노선의 상행역과 하행역으로 경로를 조회하면
      * Then 3개의 역이 출력된다
